% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.R
\name{suggestions}
\alias{suggestions}
\title{Given a history of words, returns word suggestions (predictions) ordered by greatest probability.}
\usage{
suggestions(model, history, n = 5, with_probabilities = FALSE)
}
\arguments{
\item{model}{the prediction model. At the moment there's a single one for the English language
(named "model"). Future releases might include new ones (e.g. "model_german" or "model_french").}

\item{history}{history of words (string) used to predict the next word (suggestion).}

\item{n}{number of suggestions to return.}

\item{with_probabilities}{if TRUE, names the suggestions with their correspondent
probabilities.}
}
\description{
The function uses an included model (a Markov transition matrix) generated with the notebook
created for this project.
}
